+++ Utilities/TestKextInject/KextInject.c
@@ -471,6 +471,14 @@ ApplyKernelPatches (
     } else {
       DEBUG ((DEBUG_WARN, "[OK] KernelQuirkPowerTimeoutKernelPanic patch\n"));
     }
+
+    Status = KernelApplyQuirk (KernelQuirkSegmentJettison, &Patcher, KernelVersion);
+    if (EFI_ERROR (Status)) {
+      DEBUG ((DEBUG_WARN, "[FAIL] KernelQuirkSegmentJettison - %r\n", Status));
+      FailedToProcess = TRUE;
+    } else {
+      DEBUG ((DEBUG_WARN, "[OK] KernelQuirkSegmentJettison patch\n"));
+    }
   } else {
     DEBUG ((DEBUG_WARN, "Failed to find kernel - %r\n", Status));
     FailedToProcess = TRUE;